backend/pyproject.toml
[tool.poetry]
name = "guidance-llm-backend"
version = "0.1.0"
description = "Backend services for the Guidance LLM application."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"

# --- API Gateway & Core ---
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.30"
geoalchemy2 = "^0.14.6"
psycopg2-binary = "^2.9.9"
pydantic = "^2.7.1"

# --- Security ---
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"

# --- Worker & Asynchronous Tasks ---
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.0.4"
watchfiles = "^0.21.0"

# --- LangChain & AI ---
langchain = "^0.2.1"
langchain-core = "^0.2.3"
langgraph = "^0.0.60"
ollama = "^0.2.1"

# --- Voice Services ---
TTS = "^0.22.0"
pyyaml = "^6.0.1"
# whisper-cpp-python = "0.4.0"
pydub = "^0.25.1"
numpy = "^1.26.4"
torch = "^2.3.0"

# --- Information Services ---
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"

# --- Geospatial Services ---
shapely = "^2.0.4"
pyproj = "^3.6.1"
geojson = "^3.1.0"

[tool.poetry.group.dev.dependencies]
# 今後、テストライブラリ(pytest)やリンター(ruff)などをここに追加していく
# 例: pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
